Import code for MVAPICH2 2.3a
git-svn-id: https://scm.mvapich.cse.ohio-state.edu/svn/mpi/mvapich2/trunk@8612 09bc9535-d30e-0410-b1f7-d46b20a4725c

+++ src/mpid/ch3/channels/mrail/src/gen2/vbuf.h
@@ -12,7 +12,7 @@
  *          Michael Welcome  <mlwelcome@lbl.gov>
  */
 
-/* Copyright (c) 2001-2016, The Ohio State University. All rights
+/* Copyright (c) 2001-2017, The Ohio State University. All rights
  * reserved.
  *
  * This file is part of the MVAPICH2 software package developed by the
@@ -168,8 +168,8 @@ typedef struct vbuf
     VBUF_FLAG_TYPE* head_flag;
     unsigned char* buffer;
 
-    int content_size;
-    int content_consumed;
+    size_t content_size;
+    size_t content_consumed;
 
     /* used to keep track of eager sends */
     uint8_t eager;
